refactor(bench): limit create-zone changes
